+++ tests/assets/ignore5.ts
@@ -1,0 +1,10 @@
+export interface MyApi {
+  /**
+   * @dets_ignore
+   */
+  foo: string;
+  /**
+   * @ignore
+   */
+  bar: boolean;
+}